+++ R/zzz_get_endpoint.R
@@ -2,7 +2,7 @@
 #'
 #' @inheritParams mermaid_GET
 #' @noRd
-get_endpoint <- function(endpoint = c("benthicattributes", "choices", "fishfamilies", "fishgenera", "fishspecies", "fishsizes", "managements", "projects", "projecttags", "sites", "summarysampleevents", "summarysites"), limit = NULL, filter = NULL, ...) {
+get_endpoint <- function(endpoint = c("benthicattributes", "choices", "fishfamilies", "fishgenera", "fishspecies", "fishsizes", "managements", "projects", "projecttags", "sites", "summarysampleevents", "classification/labelmappings"), limit = NULL, filter = NULL, ...) {
   url <- base_url
 
   endpoint <- match.arg(endpoint, several.ok = TRUE)
@@ -175,5 +175,6 @@ mermaid_endpoint_columns <- list(
   projects = projects_columns,
   projecttags = projecttags_columns,
   sites = sites_columns,
-  summarysampleevents = summary_sampleevents_columns
+  summarysampleevents = summary_sampleevents_columns,
+  "classification/labelmappings" = classification_labelmappings_columns
 )